# Migrate thnn_conv_depthwise2d from THC to ATen