--- a/tools/build_variables.bzl
+++ b/tools/build_variables.bzl
@@ -1005,7 +1005,6 @@ aten_native_source_non_codegen_list = [
     "aten/src/ATen/native/Integration.cpp",
     "aten/src/ATen/native/Itertools.cpp",
     "aten/src/ATen/native/LegacyBridge.cpp",
-    "aten/src/ATen/native/LegacyNNDefinitions.cpp",
     "aten/src/ATen/native/Lerp.cpp",
     "aten/src/ATen/native/Linear.cpp",
     "aten/src/ATen/native/LinearAlgebra.cpp",